python_service/wifi_core/views.py
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework import status
from .serializers import LoginSerializer, UserSerializer

class LoginAPIView(APIView):
    def post(self, request):
        serializer = LoginSerializer(data=request.data)
        if serializer.is_valid():
            user = serializer.validated_data
            # Di sini kita bisa tambahkan Token JWT nanti
            return Response({
                "message": "Login Berhasil!",
                "user": UserSerializer(user).data,
                "token": "dummy-token-sementara" 
            }, status=status.HTTP_200_OK)
        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
