requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
mysqlclient
